requirements.txt
# requirements.txt
pypdf
langchain-text-splitters
sentence-transformers
supabase
psycopg2-binary
python-dotenv
tqdm